pyproject.toml
[tool.poetry]
name = "pyback"
version = "0.1.0"
description = "FastAPI Clean Architecture backend (Neon PostgreSQL) for Chat Terminal"
readme = "README.md"
packages = [{ include = "pyback", from = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.6"
uvicorn = { extras = ["standard"], version = "^0.32.1" }
sqlalchemy = { extras = ["asyncio"], version = "^2.0.36" }
asyncpg = "^0.30.0"
alembic = "^1.14.0"
pydantic-settings = "^2.6.1"
bcrypt = ">=4.2.0,<6"
pyjwt = "^2.10.1"
groq = "^0.13.1"
cryptography = "^44.0.0"
python-multipart = "^0.0.20"
greenlet = "^3.3.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-asyncio = "^0.24.0"
ruff = "^0.8.4"
httpx = "^0.28.1"
psycopg2-binary = "^2.9.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 100
target-version = "py311"
